Don't use the whole icmp sequence # block strategy.
It turns out the identifier in the type 0/8 icmp and NOT the sequence
number is what's being changed. The identifier == tcp source port so
only need to set it once like tcp source port.

Fix sr_nat_insert_mapping not actually marking an id as taken once it
finds an empty one. Also sr_nat_timeout to do aux_ext -1024 to get the
spot in the array instead of another +1024. Useable external ports go
from 1024-65535 but the array goes from 0-64XXX. This means once you
mark a spot in the array as true, you +1024 to get the real aux_ext.
Consequently you -1024 to get the spot in the array from the aux_ext.