Store everything in "THE nat" as network byte order.
All the aux_ints are going to be in network byte order because they're
pulled form the headers. All aux_exts are blindly put into the header
when doing the faking so they MUST be in network order. Much easier to
have the nat code do the mapping than vigillantly taking care in
handle_ip.